update rand to 0.9:
This contains a lot of renaming, in rand, thread_rng got changed to rng and gen got changed to random because of the upcomming rust2024 edition.
Additionally the traits are changing.
This is highly necessary for a toolchain update

+++ voxygen/src/audio/mod.rs
@@ -612,7 +612,7 @@ impl AudioFrontend {
                     .expect("Failed to determine sound file for this trigger item."),
                 _ => {
                     // If more than one file is listed, choose one at random
-                    let rand_step = rand::random::<usize>() % item.files.len();
+                    let rand_step = (rand::random::<u64>() as usize) % item.files.len();
                     &item.files[rand_step]
                 },
             };
